Rewritten the code so that the method spawning the ChromeDriver instances runs in a separate thread, with methods to update the status on the main thread.
Still some problems with SocketsExceptions when closing the application AFTER killing the ChromeDriver processes first. If this is not done (the ChromeDriver processes are killed automatically on exit), the application exits without this exception.

Also, a WebDriverException started to occur after changing some of the application exit code. Needs looking into.